expand slave_config to consume multiple configs
Different labels have different needs, in where it regards to the
slave_setup jenkins plugin.
For example ubuntu, and centos boxes requires us to manipulate the
docker config file and to inject some variables to the slave.
But OSX slaves on Cloudbees don't support any of this, so we make
sure we don't run any of those slave_setup setups on OSX labels.
This patch allows for different 'profiles' to be used according
different slave 'labels'