Don't always restart slave agents
If the slave is spinning up for the first time and the ReconnectDisconnectedSlaves
job happens to run during the boot of the slave, the job will
disconnect the slave and re-connect it again.
Lets make sure this doesn't happen by checking if the slave is still
disconnected after a 180seconds wait. Then restart the slave agent