Use a local copy of the CERN repo signing key.
When building a Jenkins master node we install mesos components
from the CERN repository.

As part of that we download their signing key and import it.

They stopped publishing that key on their site (probably an oversight).

We can use an in-branch copy of the key until they republish it.