Render acceptance.yml as JSON with escaped chars.
Prior to this commit, jinja2 was implicitly rendering acceptance.config as
<something>, and this something could be parsed by the python yaml library.
This worked fine for everything _except_ for the new private key that is going
to be added in order to do authentication with GCE.

After a bunch of fiddling around, I decided to explicitly have jinja2 render
the yaml as JSON (because it needs to fit on 1 line due to the way ansible
parses the "content" option of things using the "copy" module). And we also
needed to re-escape any backslashes because one of the layer was rendering '\n'
as newlines before acceptance.yaml was written out to the jenkins master.